+++ common/src/main/java/me/pandadev/fallingtrees/FallingTrees.java
@@ -1,12 +1,9 @@
 package me.pandadev.fallingtrees;
 
 import com.google.gson.Gson;
-import com.mojang.blaze3d.platform.InputConstants;
-import dev.architectury.event.events.client.ClientTickEvent;
 import dev.architectury.event.events.common.PlayerEvent;
 import dev.architectury.networking.NetworkManager;
 import dev.architectury.platform.Platform;
-import dev.architectury.registry.client.keymappings.KeyMappingRegistry;
 import dev.architectury.registry.client.level.entity.EntityRendererRegistry;
 import dev.architectury.registry.registries.DeferredRegister;
 import dev.architectury.registry.registries.RegistrySupplier;
@@ -14,11 +11,11 @@ import io.netty.buffer.Unpooled;
 import me.pandadev.fallingtrees.client.renderer.TreeRenderer;
 import me.pandadev.fallingtrees.entity.TreeEntity;
 import me.pandadev.fallingtrees.network.PacketHandler;
+import me.pandadev.fallingtrees.registries.Keybindings;
 import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.ConfigHolder;
 import me.shedaniel.autoconfig.serializer.GsonConfigSerializer;
 import net.fabricmc.api.EnvType;
-import net.minecraft.client.KeyMapping;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.FriendlyByteBuf;
@@ -52,32 +49,16 @@ public class FallingTrees {
 			EntityType.Builder.of(TreeEntity::new, MobCategory.MISC).sized(1f, 1f)
 					.fireImmune().build("tree"));
 
-	public static final KeyMapping SINGLE_BLOCK_MINING_TOGGLE = new KeyMapping(
-		"key.fallingtrees.single_block_toggle_key",
-		InputConstants.Type.KEYSYM,
-		InputConstants.KEY_N,
-		KeyMapping.CATEGORY_GAMEPLAY
-	);
-
 	public static void init() {
 		AutoConfig.register(FallingTreesConfig.class, GsonConfigSerializer::new);
 		configHolder = AutoConfig.getConfigHolder(FallingTreesConfig.class);
 		serverConfig = configHolder.getConfig();
 
-		KeyMappingRegistry.register(SINGLE_BLOCK_MINING_TOGGLE);
-
 		SOUNDS.register();
 		ENTITIES.register();
 		if (Platform.getEnv() == EnvType.CLIENT) {
 			clientInit();
-
-			ClientTickEvent.CLIENT_POST.register(minecraft -> {
-				while (configHolder.getConfig().one_block_mining_method.equals(FallingTreesConfig.OneBlockMiningEnum.KEYBIND_TOGGLE) &&
-						SINGLE_BLOCK_MINING_TOGGLE.consumeClick()) {
-					configHolder.getConfig().is_mining_one_block = !configHolder.getConfig().is_mining_one_block;
-					configHolder.save();
-				}
-			});
+			Keybindings.init();
 		}
 
 		PlayerEvent.PLAYER_JOIN.register(FallingTrees::onPlayerJoin);