+++ Changelog.md
@@ -1,3 +1,9 @@
-## Version 0.5.1 Changelog
+## Version 0.5.3 Changelog
 * **Fixes**
-  * The mod trying to access a private variable
+  * Crash when chopping trees on the server.
+  * Mining speed staying slow even when tree size limit is reached.
+
+
+* **Changes**
+  * Default tree block amount limit is change from 64 to 200
+  * Sound effect settings moved to server side because it did not work on client.